# docs/openapi.json
{
	"openapi": "3.0.0",
	"paths": {
		"/trips": {
			"post": {
				"operationId": "create",
				"summary": "Create a new trip",
				"description": "Create a new trip with the provided participants.",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTripDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Trip successfully created.",
						"headers": {
							"Location": {
								"description": "URL of the created trip",
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input"
					}
				},
				"tags": ["Trips"]
			}
		},
		"/trips/{id}": {
			"get": {
				"operationId": "find",
				"summary": "Retrieve a trip",
				"description": "Gets the details of a trip by its ID.",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Trip successfully retrieved.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FindTripDto"
								}
							}
						}
					},
					"404": {
						"description": "Trip not found"
					}
				},
				"tags": ["Trips"]
			},
			"patch": {
				"operationId": "update",
				"summary": "Update a trip",
				"description": "Updates an existing tip with new details.",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTripDto"
							}
						}
					}
				},
				"responses": {
					"204": {
						"description": "Trip successfully updated.",
						"headers": {
							"Location": {
								"description": "URL of the updated trip",
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"404": {
						"description": "Trip not found"
					}
				},
				"tags": ["Trips"]
			},
			"delete": {
				"operationId": "remove",
				"summary": "Delete a trip",
				"description": "Deletes a trip by its ID.",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Trip successfully deleted."
					},
					"404": {
						"description": "Trip not found"
					}
				},
				"tags": ["Trips"]
			}
		},
		"/trips/{id}/expenses": {
			"post": {
				"operationId": "createExpense",
				"summary": "Create an expense",
				"description": "Creates an expense for a trip.",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Expense successfully created."
					},
					"400": {
						"description": "Invalid input"
					},
					"404": {
						"description": "Trip not found"
					}
				},
				"tags": ["Trips"]
			}
		},
		"/": {
			"get": {
				"operationId": "getHello",
				"parameters": [],
				"responses": {
					"200": {
						"description": ""
					}
				}
			}
		}
	},
	"info": {
		"title": "Trip Settle",
		"description": "The Trip Settle API Documentation",
		"version": "1.0",
		"contact": {}
	},
	"tags": [],
	"servers": [],
	"components": {
		"schemas": {
			"CreateTripDto": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string",
						"description": "The title of the trip",
						"minLength": 1,
						"example": "Summer Vacation"
					},
					"description": {
						"type": "string",
						"description": "The description of the trip",
						"example": "A trip to the beach with friends."
					},
					"participants": {
						"description": "Participants of the trip",
						"example": ["Alice", "Bob", "Charlie"],
						"uniqueItems": true,
						"type": "array",
						"items": {
							"type": "string",
							"minLength": 1
						}
					}
				},
				"required": ["title", "description", "participants"]
			},
			"ExpenseDto": {
				"type": "object",
				"properties": {
					"amount": {
						"type": "number",
						"description": "The amount of money spent for this particular expense",
						"minimum": 0,
						"exclusiveMinimum": true,
						"example": 50
					},
					"currency": {
						"type": "string",
						"description": "The currency in which the expense is recorded, following the ISO 4217 standard",
						"minLength": 3,
						"maxLength": 3,
						"pattern": "^[A-Z]{3}$",
						"example": "USD"
					},
					"description": {
						"type": "string",
						"description": "A short description of what the expense was for",
						"minLength": 1,
						"maxLength": 255,
						"example": "Dinner at a restaurant."
					},
					"note": {
						"type": "string",
						"description": "An optional detailed note providing more information about the expense",
						"maxLength": 1000,
						"example": "10% tip included."
					},
					"participants": {
						"description": "A list of participants involved in the expense, who will divide the cost",
						"minItems": 1,
						"uniqueItems": true,
						"example": ["Alice", "Bob", "Charlie"],
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"payer": {
						"type": "string",
						"description": "The person who paid for the expense",
						"minLength": 1,
						"maxLength": 255,
						"example": "Alice"
					}
				},
				"required": ["amount", "currency", "description", "note", "participants", "payer"]
			},
			"FindTripDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"description": "Unique Base64URL format ID of the trip"
					},
					"title": {
						"type": "string",
						"description": "The title of the trip",
						"minLength": 1,
						"maxLength": 100,
						"example": "Summer Vacation"
					},
					"description": {
						"type": "string",
						"description": "The description of the trip",
						"minLength": 1,
						"maxLength": 500,
						"example": "A trip to the beach with friends."
					},
					"participants": {
						"description": "Array of all participants in the trip",
						"example": ["Alice", "Bob", "Charlie"],
						"uniqueItems": true,
						"type": "array",
						"items": {
							"type": "string",
							"minLength": 1
						}
					},
					"expenses": {
						"description": "Array of expenses associated with the trip",
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ExpenseDto"
						}
					}
				},
				"required": ["id", "title", "description", "participants", "expenses"]
			},
			"UpdateTripDto": {
				"type": "object",
				"properties": {}
			}
		}
	}
}